Jukebox cpp, h, and namespace splitting from Audio
