+++ src/OpenLoco/CMakeLists.txt
@@ -163,6 +163,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Input/Shortcuts.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Interop/Hooks.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Intro.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Jukebox.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Localisation/Conversion.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Localisation/Formatting.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Localisation/LanguageFiles.cpp"
@@ -509,6 +510,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Input/Shortcuts.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Interop/Hooks.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Intro.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Jukebox.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/LabelFrame.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Localisation/Conversion.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Localisation/FormatArguments.hpp"
